feat: remote execution is now behaving better
