+++ src/Xenial.Tasty/Execution/TestRuntime/DetectInteractiveRunMiddleware.cs
@@ -1,0 +1,18 @@
+﻿namespace Xenial.Delicious.Execution.TestRuntime
+{
+    public static class DetectInteractiveRunMiddleware
+    {
+        public static TestExecutor UseInteractiveRunDetection(this TestExecutor executor)
+            => executor.UseRuntime(async (context, next) =>
+            {
+                try
+                {
+                    context.IsInteractive = await context.Scope.IsInteractiveRunHook();
+                }
+                finally
+                {
+                    await next();
+                }
+            });
+    }
+}
