+++ src/Xenial.Tasty/Execution/TestRuntime/RemoteStreamDisposalMiddleware.cs
@@ -1,0 +1,22 @@
+﻿namespace Xenial.Delicious.Execution.TestRuntime
+{
+    public static class RemoteStreamDisposalMiddleware
+    {
+        public static TestExecutor UseRemoteDisposal(this TestExecutor executor)
+            => executor.UseRuntime(async (context, next) =>
+            {
+                try
+                {
+                    await next();
+                }
+                finally
+                {
+                    if (!context.IsInteractive)
+                    {
+                        context.RemoteStream?.Dispose();
+                        context.Remote?.Dispose();
+                    }
+                }
+            });
+    }
+}
