+++ src/Xenial.Tasty/Execution/TestRuntime/ResetConsoleColorMiddleware.cs
@@ -1,0 +1,20 @@
+﻿using System;
+
+namespace Xenial.Delicious.Execution.TestRuntime
+{
+    public static class ResetConsoleColorMiddleware
+    {
+        public static TestExecutor UseResetConsoleColor(this TestExecutor executor)
+            => executor.UseRuntime(async (context, next) =>
+            {
+                try
+                {
+                    await next();
+                }
+                finally
+                {
+                    Console.ResetColor();
+                }
+            });
+    }
+}
